Refactor Package Structure
