+++ GameEngine/src/main/java/org/euph/engine/SceneLoadMode.java
@@ -1,4 +1,4 @@
-package org.euph.engine.entityComponentSystem;
+package org.euph.engine;
 
 public enum SceneLoadMode {
     SINGLE,
