--- GameEngine/src/main/java/org/euph/engine/entityComponentSystem/systems/render/RenderSystem.java
@@ -1,7 +1,0 @@
-package org.euph.engine.entityComponentSystem.systems.render;
-
-//TODO: this
-//TODO: JavaDoc
-public class RenderSystem {
-
-}
