--- GameEngine/src/main/java/org/euph/engine/entityComponentSystem/systems/render/pipelines/RenderPipeline3d.java
@@ -1,6 +1,0 @@
-package org.euph.engine.entityComponentSystem.systems.render.pipelines;
-
-import org.euph.engine.entityComponentSystem.systems.render.RenderPipeline;
-
-public class RenderPipeline3d extends RenderPipeline {
-}
