+++ GameEngine/src/main/java/org/euph/engine/systems/render/RenderSystem.java
@@ -1,0 +1,7 @@
+package org.euph.engine.systems.render;
+
+//TODO: this
+//TODO: JavaDoc
+public class RenderSystem {
+
+}
