+++ GameEngine/src/main/java/org/euph/engine/systems/render/pipelines/RenderPipeline3d.java
@@ -1,0 +1,6 @@
+package org.euph.engine.systems.render.pipelines;
+
+import org.euph.engine.systems.render.RenderPipeline;
+
+public class RenderPipeline3d extends RenderPipeline {
+}
